Move snr_pseudo to quality assessment
